Client side authentication level control

+++ WebApplication/public/script.js
@@ -131,18 +131,25 @@ function loadFacilities_administrator(sport_center_id){
 //@param[sport_center_id]: id of the sport center where the sport facility is
 function deleteSportFacility(id_sport_facility, sport_center_id){
     var token = "empty";
+    var auth_level = "empty";
     token = getCookie("token");
-    fetch('../api/v1/sport_facilities/'+id_sport_facility, {
-        method: 'DELETE',
-        headers: { "x-access-token": token },
-    })
-    .then((resp) => {
-        if(resp.status==403){
-            console.log("Authentication error");
-        }else{
-            loadFacilities_administrator(sport_center_id);
-        }
-    })
+    auth_level = getCookie("user_level");
+
+    if(auth_level=="administrator"){
+        fetch('../api/v1/sport_facilities/'+id_sport_facility, {
+            method: 'DELETE',
+            headers: { "x-access-token": token },
+        })
+        .then((resp) => {
+            if(resp.status==403){
+                console.log("Authentication error");
+            }else{
+                loadFacilities_administrator(sport_center_id);
+            }
+        })
+    }else{
+        console.log("Authentication error");
+    }
 }
 //...
